remove the Arbitrator trait
